iota_node_storage

Trait GrpcStateReader

pub trait GrpcStateReader:
    ObjectStore
    + ReadStore
    + Send
    + Sync {
    // Required methods
    fn get_lowest_available_checkpoint_objects(
        &self,
    ) -> Result<CheckpointSequenceNumber>;
    fn get_chain_identifier(&self) -> Result<ChainIdentifier>;
    fn get_epoch_last_checkpoint(
        &self,
        epoch_id: EpochId,
    ) -> Result<Option<VerifiedCheckpoint>>;
    fn grpc_indexes(&self) -> Option<&dyn GrpcIndexes>;
    fn get_struct_layout(
        &self,
        type_tag: &StructTag,
    ) -> Result<Option<MoveTypeLayout>>;

    // Provided method
    fn get_type_layout(
        &self,
        type_tag: &TypeTag,
    ) -> Result<Option<MoveTypeLayout>> { ... }
}

Trait extending ReadStore with full-node-specific queries that may require richer databases or indexes to support.

Required Methods

fn get_lowest_available_checkpoint_objects( &self, ) -> Result<CheckpointSequenceNumber>

Lowest available checkpoint for which object data can be requested.

Specifically this is the lowest checkpoint for which input/output object data will be available.

fn get_chain_identifier(&self) -> Result<ChainIdentifier>

fn get_epoch_last_checkpoint( &self, epoch_id: EpochId, ) -> Result<Option<VerifiedCheckpoint>>

fn grpc_indexes(&self) -> Option<&dyn GrpcIndexes>

Get a handle to an instance of the gRPC indexes.

fn get_struct_layout( &self, type_tag: &StructTag, ) -> Result<Option<MoveTypeLayout>>

Provided Methods

fn get_type_layout(&self, type_tag: &TypeTag) -> Result<Option<MoveTypeLayout>>